feat: Restruct find_task via (pid, task) map
Adjust code according to clippy

+++ os/src/loader.rs
@@ -35,7 +35,7 @@ lazy_static! {
         unsafe {
             for _ in 0..num_app {
                 let mut end = start;
-                while end.read_volatile() != b'\0' as u8 {
+                while end.read_volatile() != b'\0' {
                     end = end.add(1);
                 }
                 let slice = core::slice::from_raw_parts(start, end as usize - start as usize);
